Improve styling in country page
