+++ src/views/CountryNsOverviewStrategicPriorities/i18n.json
@@ -1,10 +1,8 @@
 {
     "namespace": "countryNsOverviewStrategicPriorities",
     "strings": {
-        "countryPlanTitle": "{countryName} Country Plan",
-        "countryPlanNoCountryPlan": "This country hasn't published any Country Plan yet!",
+        "nsStrategicPrioritiesHeading": "NS Strategic Priorities",
         "countryPlanPeopleTargeted": "People Targeted",
-        "countryPlanLoadFailureMessage": "Failed to load the Country Plan!",
         "countryPlanDownloadPlan": "{countryName} Country Plan",
         "countryPlanDownloadPlanInternal": "{countryName} Internal Plan",
         "countryPlanKeyFigureRequestedAmount": "Federation-Wide Requested Amount (CHF)"
