+++ src/views/CountryNsOverviewStrategicPriorities/index.tsx
@@ -6,11 +6,10 @@ import { type CountryOutletContext } from '#utils/outletContext';
 import KeyFigure from '#components/KeyFigure';
 import Link from '#components/Link';
 import BlockLoading from '#components/BlockLoading';
-import Message from '#components/Message';
 import Container from '#components/Container';
+import useTranslation from '#hooks/useTranslation';
 import { resolveToString } from '#utils/translation';
 import { useRequest } from '#utils/restRequest';
-import useTranslation from '#hooks/useTranslation';
 
 import StrategicPrioritiesTable from './StrategicPrioritiesTable';
 
@@ -26,7 +25,7 @@ export function Component() {
         pending: countryPlanPending,
         response: countryPlanResponse,
     } = useRequest({
-        skip: !countryId || isNotDefined(countryId) || !countryResponse?.has_country_plan,
+        skip: isNotDefined(countryId) || !countryResponse?.has_country_plan,
         url: '/api/v2/country-plan/{country}/',
         pathVariables: {
             country: Number(countryId),
@@ -35,74 +34,60 @@ export function Component() {
 
     return (
         <Container
-            className={styles.countryPlan}
-            heading={resolveToString(
-                strings.countryPlanTitle,
-                { countryName: countryResponse?.name ?? '--' },
-            )}
-            headingLevel={2}
+            className={styles.countryNsOverviewStrategicPriorities}
+            heading={strings.nsStrategicPrioritiesHeading}
             withHeaderBorder
         >
-            {isDefined(countryPlanResponse) && (
-                <div className={styles.countryPlanDownloadLink}>
-                    {isDefined(countryPlanResponse.public_plan_file) && (
-                        <Link
-                            variant="secondary"
-                            href={countryPlanResponse.public_plan_file}
-                            external
-                            className={styles.downloadLink}
-                            icons={<DownloadLineIcon className={styles.icon} />}
-                        >
-                            {resolveToString(
-                                strings.countryPlanDownloadPlan,
-                                { countryName: countryResponse?.name ?? '--' },
-                            )}
-                        </Link>
-                    )}
-                    {isTruthyString(countryPlanResponse.internal_plan_file) && (
-                        <Link
-                            variant="secondary"
-                            href={countryPlanResponse.internal_plan_file}
-                            external
-                            className={styles.downloadLink}
-                            icons={<DownloadLineIcon className={styles.icon} />}
-                        >
-                            {resolveToString(
-                                strings.countryPlanDownloadPlanInternal,
-                                { countryName: countryResponse?.name ?? '--' },
-                            )}
-                        </Link>
-                    )}
-                </div>
-            )}
-            {!countryResponse?.has_country_plan && (
-                <Message
-                    title={strings.countryPlanNoCountryPlan}
-                />
-            )}
             {countryPlanPending && (
                 <BlockLoading />
             )}
-            {countryResponse?.has_country_plan && !countryPlanPending && !countryPlanResponse && (
-                <div className={styles.errored}>
-                    {strings.countryPlanLoadFailureMessage}
-                </div>
-            )}
-            {countryResponse?.has_country_plan && !countryPlanPending && countryPlanResponse && (
-                <div className={styles.content}>
-                    <div className={styles.keyFigures}>
-                        <KeyFigure
-                            className={styles.keyFigure}
-                            value={countryPlanResponse.requested_amount}
-                            label={strings.countryPlanKeyFigureRequestedAmount}
-                            compactValue
-                        />
-                        <KeyFigure
-                            className={styles.keyFigure}
-                            value={countryPlanResponse.people_targeted}
-                            label={strings.countryPlanPeopleTargeted}
-                            compactValue
-                        />
+            {isDefined(countryPlanResponse) && (
+                <div className={styles.countryPlan}>
+                    <div className={styles.downloadLinksAndKeyFigures}>
+                        <div className={styles.countryPlanDownloadLink}>
+                            {isDefined(countryPlanResponse.public_plan_file) && (
+                                <Link
+                                    variant="secondary"
+                                    href={countryPlanResponse.public_plan_file}
+                                    external
+                                    className={styles.downloadLink}
+                                    icons={<DownloadLineIcon className={styles.icon} />}
+                                >
+                                    {resolveToString(
+                                        strings.countryPlanDownloadPlan,
+                                        { countryName: countryResponse?.name ?? '--' },
+                                    )}
+                                </Link>
+                            )}
+                            {isTruthyString(countryPlanResponse.internal_plan_file) && (
+                                <Link
+                                    variant="secondary"
+                                    href={countryPlanResponse.internal_plan_file}
+                                    external
+                                    className={styles.downloadLink}
+                                    icons={<DownloadLineIcon className={styles.icon} />}
+                                >
+                                    {resolveToString(
+                                        strings.countryPlanDownloadPlanInternal,
+                                        { countryName: countryResponse?.name ?? '--' },
+                                    )}
+                                </Link>
+                            )}
+                        </div>
+                        <div className={styles.keyFigures}>
+                            <KeyFigure
+                                className={styles.keyFigure}
+                                value={countryPlanResponse.requested_amount}
+                                label={strings.countryPlanKeyFigureRequestedAmount}
+                                compactValue
+                            />
+                            <KeyFigure
+                                className={styles.keyFigure}
+                                value={countryPlanResponse.people_targeted}
+                                label={strings.countryPlanPeopleTargeted}
+                                compactValue
+                            />
+                        </div>
                     </div>
                     <StrategicPrioritiesTable
                         className={styles.strategicPriorityTable}
