+++ src/views/CountryProfileSupportingPartners/styles.module.css
@@ -1,11 +1,6 @@
-.membership-coordination-table {
-    .icon {
-        color: var(--go-ui-color-primary-red);
-        font-size: var(--go-ui-height-icon-multiplier);
-    }
-
-    .sector-heading {
-        transform: rotate(180deg);
-        writing-mode: vertical-lr;
-    }
+.country-profile-supporting-partners {
+    display: flex;
+    flex-direction: column;
+    gap: var(--go-ui-spacing-2xl);
+    padding: var(--go-ui-spacing-2xl) 0;
 }
